requirements.txt
# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# AI/ML
llama-index>=0.10.0
llama-index-llms-google-genai
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-parse

# Visualização
plotly>=5.15.0
altair>=5.0.0

# Utilitários
python-dotenv>=1.0.0
deep-translator>=1.11.0
openpyxl>=3.0.0
pypdf>=3.0.0

# Desenvolvimento
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0